requirements.txt
# Core scientific
numpy>=1.23,<2.0
pandas>=1.5,<2.0
scipy>=1.9,<2.0

# Plotting
matplotlib>=3.6,<4.0
seaborn>=0.12,<0.13

# Utilities
pyyaml>=6.0,<7.0
tqdm>=4.64,<5.0
tzdata>=2023.3
pytmatrix